+++ use_class.py
@@ -14,13 +14,8 @@ pw = getpass.getpass()
 hvconnectionobj = vmware_horizon.Connection(username = username,domain = domain,password = pw,url = url)
 hvconnectionobj.hv_connect()
 #print(hvconnectionobj)
-obj = vmware_horizon.External(url=hvconnectionobj.url, access_token=hvconnectionobj.access_token)
-monitor = vmware_horizon.Monitor(url=hvconnectionobj.url, access_token=hvconnectionobj.access_token)
-vc = (monitor.virtual_centers())[0]
-print(vc['id'])
-dc = obj.get_datacenters(vcenter_id = vc["id"])[0]
-hoc = obj.get_hosts_or_clusters(vcenter_id = vc["id"], datacenter_id=dc["id"])
-print(hoc)
+obj = vmware_horizon.Inventory(url=hvconnectionobj.url, access_token=hvconnectionobj.access_token)
+sessions = obj.get_sessions(maxpagesize=1)
 
 
 # print(dc)